# infra/swagger.yaml
openapi: 3.0.0
info:
  title: Quark Platform API
  version: 1.0.0
  description: |
    Централизованная документация всех API endpoints платформы Quark MKS.
    
    ## Аутентификация
    Большинство endpoints требуют JWT токен в заголовке Authorization:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Базовые URL
    - **API Gateway**: `http://api.quark.local` (через Traefik)
    - **Monitoring**: `http://monitor.quark.local`
    - **Admin UI**: `http://admin.quark.local`
    
  contact:
    name: Quark Platform Team
    email: support@quark.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://api.quark.local
    description: Production API Gateway (через Traefik)
  - url: http://localhost
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен, полученный через /auth/login
  schemas:
...
    UserRegistration:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 6
          example: SecurePass123!
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        phone:
          type: string
          example: "+1234567890"
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: SecurePass123!
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          example: 8e2eb370-69d2-42f7-919c-8f15cb1d794d
        email:
          type: string
          example: user@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        roles:
          type: array
          items:
            type: string
          example: ["user", "admin"]
        isEmailVerified:
          type: boolean
          example: true

    # Blog schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Invalid input data"
        statusCode:
          type: integer
          example: 400
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      description: Создание нового пользователя в системе
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Авторизация пользователя
      description: Получение JWT токена для доступа к защищенным endpoints
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...